split teams views for oss and enterprise
